fix #7028: API to create integration test pages
It enabled clients to control how the application is started and do any additional authentication if it is required.
